rova: patches: Fix sometimes not applied patches (git WTF?!)
Sometimes git silently fails with exit code 0 and no error output. But
it outputs 'Patch skipped' when using --verbose arg. People say this
happens when patching a non-git directory, but from my experience this
isn't true and 99% of the time it worked. Git should fall back to the
"patch" command when patching outside of git repo, but for some
patches it doesn't happen, so let's switch completely to "patch" to
make sure this never repeat again.

Signed-off-by: iusmac <iusico.maxim@libero.it>